database are now accessible without separate download
Codes:
' Get the path to the application directory
Dim databasePath As String = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "disc.db")
Dim connectionString As String = $"Data Source={databasePath};Version=3;"

Place disc.db in the Output Directory:

Move disc.db to your project's root folder (or wherever you plan to keep it relative to the .exe).
In Solution Explorer, right-click disc.db and go to Properties:
Set Copy to Output Directory to Copy always or Copy if newer.
Deploy Your Application:

Ensure that disc.db is included in the same directory as your .exe or packaged with your application if you're using an installer.
Verify After Deployment:

When you deploy or copy the application to another computer, make sure the database file (disc.db) is in the same relative location (next to or in a subfolder of your .exe).